% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{point2Frame}
\alias{point2Frame}
\title{Make a halton frame around a BAS point}
\usage{
point2Frame(
  pt,
  bb = st_bbox(st_as_sf(data.frame(t(bb)), coords = c("x", "y"), crs = nztm)),
  base = c(2, 3),
  J = c(2, 2),
  projstring = NULL
)
}
\description{
Make a halton frame around a BAS point
}
